Removed mode-change capability.
Kernel (probably generic hiddriver) manages the device already in
tablet mode. The position sensing in mouse mode is quite noisy, I'm
not sure if it is needed at all. Tablet mode already works and
dual-mode capability might be a bit confusing for users.

The mode-change command is implemented in m210.pegatech.M210, but it
is hidden from user interface for now.